+++ KeyboardKit/Gestures/UIView+TapAction.swift
@@ -11,7 +11,7 @@
 /*
  
  This extension applies tap gesture recognizers using action
- blocks instead of the target/selector pattern.
+ blocks instead of a target and a selector.
  
  */
 
@@ -19,20 +19,20 @@ import UIKit
 
 public extension UIView {
     
-    typealias TapAction = (() -> Void)
+    typealias TapAction = () -> Void
     
-    func addTapAction(numberOfTapsRequired: Int = 1, action: TapAction?) {
+    func addTapAction(numberOfTapsRequired: Int = 1, action: @escaping TapAction) {
         tapAction = action
         isUserInteractionEnabled = true
-        let recognizer = UITapGestureRecognizer(target: self, action: #selector(handleTap))
+        let recognizer = UITapGestureRecognizer(target: self, action: #selector(handleTapAction))
         recognizer.numberOfTapsRequired = numberOfTapsRequired
         addGestureRecognizer(recognizer)
     }
     
     func removeTapAction() {
-        guard let gestures = gestureRecognizers else { return }
-        let taps = gestures.filter { $0 is UITapGestureRecognizer }
-        taps.forEach { removeGestureRecognizer($0) }
+        gestureRecognizers?
+            .filter { $0 is UITapGestureRecognizer }
+            .forEach { removeGestureRecognizer($0) }
     }
 }
 
@@ -51,7 +51,7 @@ private extension UIView {
         }
     }
 
-    @objc func handleTap(sender: UITapGestureRecognizer) {
+    @objc func handleTapAction(sender: UITapGestureRecognizer) {
         tapAction?()
     }
 }