+++ KeyboardKit/Views/Toolbars/KeyboardToolbar.swift
@@ -8,9 +8,12 @@
 
 /*
  
- This class behaves like `KeyboardButtonRow`, but instead of
- providing it with actions and a button creation block, this
- class lets you add any views you like to its `stackView`.
+ Keyboard toolbars are "toolbar rows" that can be added to a
+ `KeyboardInputViewController`s `keyboardStackView`.
+ 
+ This is a base class, that you can inherit to create custom
+ toolbars. It has a horizontal `stackView`, to which you can
+ add any views you like.
  
  */
 
@@ -19,7 +22,7 @@ import UIKit
 open class KeyboardToolbar: UIView, KeyboardStackViewComponent {
     
     public convenience init(
-        height: CGFloat,
+        height: CGFloat = .standardKeyboardRowHeight,
         alignment: UIStackView.Alignment = .fill,
         distribution: UIStackView.Distribution = .fillEqually) {
         self.init(frame: .zero)