+++ KeyboardKitExampleKeyboard/DemoKeyboardActionHandler.swift
@@ -6,32 +6,62 @@
 //  Copyright © 2019 Daniel Saidi. All rights reserved.
 //
 
-/*
+import KeyboardKit
+
+/**
  
  This action handler inherits `StandardKeyboardActionHandler`
  and adds demo-specific functionality to it.
  
  */
-
-import KeyboardKit
-
 class DemoKeyboardActionHandler: StandardKeyboardActionHandler {
+    
+    
+    // MARK: - Initialization
+    
+    public init(inputViewController: UIInputViewController) {
+        super.init(
+            inputViewController: inputViewController,
+            tapHapticFeedback: .standardTapFeedback,
+            longPressHapticFeedback: .standardLongPressFeedback
+        )
+    }
+    
+    
+    // MARK: - Properties
+    
+    var keyboardShiftState = KeyboardShiftState.lowercased
     
     var demoViewController: KeyboardViewController? {
         return inputViewController as? KeyboardViewController
     }
     
+    
+    // MARK: - Functions
+    
+    func animateButtonTap(for view: UIView) {
+        (view as? KeyboardButton)?.animateStandardTap()
+    }
+    
     override func handleLongPress(on action: KeyboardAction, view: UIView) {
-        super.handleLongPress(on: action, view: view)
-        saveImage(for: action)
+        animateButtonTap(for: view)
+        switch action {
+        case .shift: switchToAlphabeticKeyboard(.capsLocked)
+        case .image: saveImage(for: action)
+        default: super.handleLongPress(on: action, view: view)
+        }
     }
     
     override func handleTap(on action: KeyboardAction, view: UIView) {
-        (view as? KeyboardButton)?.animateStandardTap()
+        animateButtonTap(for: view)
         super.handleTap(on: action, view: view)
         switch action {
-        case .shift: demoViewController?.keyboardType = .alphabetic(uppercased: true)
-        case .shiftDown: demoViewController?.keyboardType = .alphabetic(uppercased: false)
+        case .shift: switchToAlphabeticKeyboard(.uppercased)
+        case .shiftDown: switchToAlphabeticKeyboard(.lowercased)
+        case .character:
+            if keyboardShiftState == .uppercased {
+                switchToAlphabeticKeyboard(.lowercased)
+            }
         case .switchToKeyboard(let type): demoViewController?.keyboardType = type
         default: copyImage(for: action)
         }
@@ -48,6 +78,9 @@ class DemoKeyboardActionHandler: StandardKeyboardActionHandler {
         else { alert("Failed!") }
     }
 }
+
+
+// MARK: - Private Extensions
 
 private extension DemoKeyboardActionHandler {
     
@@ -77,5 +110,10 @@ private extension DemoKeyboardActionHandler {
         guard input.hasFullAccess else { return alert("You must enable full access to save images to photos.") }
         let saveCompletion = #selector(handleImage(_:didFinishSavingWithError:contextInfo:))
         image.saveToPhotos(completionTarget: self, completionSelector: saveCompletion)
+    }
+    
+    func switchToAlphabeticKeyboard(_ state: KeyboardShiftState) {
+        keyboardShiftState = state
+        demoViewController?.keyboardType = .alphabetic(uppercased: state.isUppercased)
     }
 }